Fix for compiler warning.

+++ src/PartFile.cpp
@@ -3974,7 +3974,8 @@ CUpDownClient* CPartFile::GetSlowerDownloadingClient(uint32 speed, CUpDownClient
 	for( SourceSet::iterator it = m_SrcList.begin(); it != m_SrcList.end(); ) {
 		CUpDownClient* cur_src = *it++;
 		if ((cur_src->GetDownloadState() == DS_DOWNLOADING) && (cur_src != caller)) {
-			uint32 factored_bytes_per_second = (cur_src->GetKBpsDown() * 1024) * DROP_FACTOR;
+			uint32 factored_bytes_per_second = static_cast<uint32>(
+				(cur_src->GetKBpsDown() * 1024) * DROP_FACTOR);
 			if ( factored_bytes_per_second< speed) {
 				printf("Selecting source %p to drop: %d < %d\n", cur_src, factored_bytes_per_second, speed);
 				printf("End slower source calculation\n");
